Extract RealmServer class methods into per-concern handler modules
server.ts went from 1242 lines to 367. Five orphaned "fake handler"
methods that lived on the class because they closed over private
state are now free functions in handlers/ + lib/realm-routing.ts,
each with a typed deps interface scoped to what it actually needs.

  handlers/create-realm.ts   — createRealm impl + the Koa middleware
                               (merged from the old handle-create-realm.ts)
  handlers/send-event.ts     — sendEvent factory
  handlers/serve-from-realm.ts — serveFromRealm factory
  handlers/serve-index.ts    — serveIndex / serveHostApp factory
                               (closes over the index-html cache that
                               previously lived as class state)
  lib/realm-routing.ts       — findOrMountRealm + getPublishedRealmInfo
                               + isIndexedCardInstance +
                               hasExtensionlessSourceModule +
                               hasPublicPermissions + candidateRealmURLs

No behaviour change. CreateRoutesArgs loses createRealm; the
realm-creation route now wires a CreateRealmDeps bag directly to the
extracted factory. testingOnlyFindOrMountRealm calls the lib helper.
